Migrated EventEmitter mixin to class.
